Refactored all JSP and servlets related to view all and add

+++ src/main/java/com/beck/javaiii_kirkwood/personal_project/controllers/DeletePrivilegeServlet.java
@@ -10,19 +10,42 @@ import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
 
 import java.io.IOException;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @WebServlet("/deleteprivilege")
 public class DeletePrivilegeServlet extends HttpServlet {
   @Override
   protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+    Map<String, String> results = new HashMap<>();
     HttpSession session = req.getSession();
     session.setAttribute("currentPage",req.getRequestURL());
     req.setAttribute("pageTitle", "Privacy Policy");
-    Integer PrivilegeID = Integer.valueOf(req.getParameter("privilegeid"));
-    int result = PrivilegeDAO.deletePrivilege(PrivilegeID);
+    int PrivilegeID = Integer.valueOf(req.getParameter("privilegeid"));
+    int mode = Integer.valueOf(req.getParameter("mode"));
+    int result = 0;
+    if (mode==0){
+      try{
+        result = PrivilegeDAO.deletePrivilege(PrivilegeID);
+      }
+      catch(Exception ex){
+        results.put("dbStatus",ex.getMessage());
+      }
+    }
+    else {
+      try{
+        result = PrivilegeDAO.undeletePrivilege(PrivilegeID);
+      }
+      catch(Exception ex){
+        results.put("dbStatus",ex.getMessage());
+      }
+
+    }
+
     List<Privilege> privileges = null;
     privileges = PrivilegeDAO.getAllPrivilege();
+    req.setAttribute("results",results);
     req.setAttribute("Privileges", privileges);
     req.setAttribute("pageTitle", "All Privileges");
     int x = 0;
